Fix Hamiltonian, improve CLI for ground state search.
As per discussion with Betre on 2020-10-17, the mins of a hamiltonian may not be 0.
The previous implementation of the simulator assumed this was the case.
Therefore, the simulator can not signal to the user that it is "done".

During our conversation, we discovered that my Hamiltonian was wrong.
I had erroniously switched the order of exponention and summation.