Implement Markov Decision Process.
Based on my discussion with Dr. B on 2020-10-24, I have implemented a MDP agent.
It isnt strictly an MDP, because it has two-state lookback.
The agent check if its last action increased or decreased the energy of the system.
If the enrgy decreased, we continue along.
If we increased the energy, that's bad.
In this case, we generate a random number, and back out our last decision based on a fn of the change in system energy.
This should improve exploration behavior by reducing the liklihood of making many "bad" flips.