Implement gradient following.
This commit implements a method that computes first deriatives of a graph.
That is, toggle every edge pair from the starting graph and record Δenergy.
A simple gradient followng algorithm would toggle the edge that minimizes the energy of the graph.
However, this approach seems to get trapped in local minima.