Omnibus: probalisitc GD, annealing, unified experiment runner.
Search strategies are not monolithic entities.
They are composed from multiple sub-strategies.
One such configuration splits "exploration" from "backtracking".
This is the formulation I created and used.

Exploration is the process of reaching to new states.
Backtracking is the process of undoing and old transition.

Metropolis process (and annealing) are forms of backtracking or conditionally accepting a state.

Exploration can then be gradient based or based on a truly random graph walk.
Gradient descent gets stuck, and benefits from some randomness, hence SGD.

Additionally, this commit changes controller.py to be a one-stop shop for all things experiment related.
Check out help to understand how it works.
Invoke it with --help or a subcommand like "run --help".