Allow comparisons on either # of toggles or # of H calls.
Grad descent is supposed to be sample efficient, which means for a given # of toggles, it should do better than other methods.
However, this is not the perfect comparison, as grad descent invokes H a bunch of times (on order of N^2).
So, add a comparison mode that judges algorithms based on the number of times they call H.