(GRAP-35) Fix step/reset on annealing strategies.
Terminology was confusing.
end_sweep was really triggered at the end of an epoch.
Thus, it has been renamed end_epoch().

Additionally, instead of auto-stepping the strategy, an explicit step() must be called.
This is ultimately triggered from our CC sampler.
This change was made so that getting an action doesn't also modify the state of the agent.