Classify based on purity ratio.
The purity ratio is (# of k-size max cliques)/(total # of max cliques).

Right now, all impure graphs are assigned the same energy (or, we try to).
That energy is 1.
However, some graphs are nearly pure, so it doesn't make sense for them to be grouped with graphs that a very impure.
So, energy should be a function of how pure a graph is.

This change refactors anything that handles a dataset, and will likely break and stage 2,3 pipelines.
We will see how this works, as it may do a terrible job at differentiating high/low energy graphs.