Common subsequence deduplication.
Defines a canonicalized list of moves that transforms a seed graph into a given output.
Moves are stored as follows:
Each move is a pair (c1, c2), where c1<c2.
Node c2 (and all of its edges) are merged into c1.
The list of moves is sorted from greatest to least on c1, and then from least to greatest on c2.

If the current sequence for a graph matches an existing subsequence, then that path will not be explored.
This method cuts down the number of graphs visited by a factor of at least 1k.

For some reason, graphs with 5 faces are being generated, and I don't yet have a diagnosis for the problem.